Syntax analysis finished
All the Syntax analysis has been finished correctly until this part.

If in the next part of the symbol table you have any problem, you can create a branch from this point to restart from a fresh checkpoint.

Also here's a script that helps you to compile de JJ file and analize all the TXT that the teacher gave us. Modify if necessary because the output rewrites itself in the "temporal.txt" file